pass mysterious --headless to QApplication arguments
With Qt 5.11, blinkpdf segfaults when using offscreen without a display.
The crash can't be reproduced with Qt 5.15.

By passing --headless, it seems there is no crash anymore on Qt 5.11.
However, this option isn't documented and is nowhere to be found in Qt source,
but it does seem to have an effect.